Implement basic nudge behavior.
This turns up an unexpected bug in wall.right2: in the upper right
hand corner, i see:

   . . M
   . . .

But it should be:

   . . M
   . M M
   . . M

The problem is that the T's grid actually extends past the edge of
the playfield:

   . . . M ~
   . . M M ~
   . . . M ~
   . . . .

This is causing a problem for xyvs-update because it doesn't know
to clip the overhanging cells. Basically, it's trying to do a full
join when it should be doing a left join.